Fix some signatures

+++ include/unk_0205B3DC.h
@@ -4,7 +4,7 @@
 #include "msgfmt.h"
 
 void sub_0205B3DC(u32 trainerId, u32 playerGender, MSGFMT *msgFmt);
-u8 sub_0205B418(u32 trainerId, u32 playerGender, u32 choice);
+int sub_0205B418(u32 trainerId, u32 playerGender, u32 choice);
 int sub_0205B464(int playerGender, int trclass);
 int sub_0205B46C(int playerGender, int trclass, int attr);
 u16 *sub_0205B4A4(HeapID heapId);
